The Line Begins To Blur
This adds in the full song for The Line Begins To Blur. It also
decreases the probability that a string will break.

Those string tension values will need to be evaluated for demonstration
purposes when showing the output of the line coverage simulation;
however, they have been made less likely for testing the song without
worrying about breaking a string.