Oneshot Line Example
This adds in a simulation of a festival setlist (they've got a short
timeslot) that tracks all of the synth memory that's used throughout the
set. As a result, they can see which memory locations are accessed and
which aren't. Those that are need to be consolidated to as few keyboards
as possible.