Full Synth Rig
A band now has a keyboardist, who can have many synths. Those different
synths have different patches on them, and different memory
implementations. We can investigate which patches are used on which
synths by seeing which patch memory locations are read and which aren't.

This can then be used to modify the synths to minimize the number
necessary for a one-off festival performance.